devel branch conveniences (revert before merge!)
This is to keep git from yelling at us about changed but uncommitted
files it the repo all the time while mucking around it the devel branch.
The finished builds are not committed here. It order to bump the release
and do an actual release build a release branch should be made. A commit
should be made with the correct version number set and `git add -f` the
two prebuilt userscripts. The that branch should be merged to master.