prepare branch to be used as an alternate release
In order to hack on using remote data sources, I am setting up this
branch to be an alternate release tree until such a time as it gets
merged to master or all hope abandoned. In either case the auto-update
URL should eventually be replaced with the master one.

Branches from this for fixes or features should be made from the
corresponding <name>_devel branch and not include the prebuilt
userscripts.